Move configuration constants to a header file and abstract the updates to the haptic points.
The constants use for configuring IO and networking were moved to the config.h file so they can be easier edited in the future.

The operation of updating the motor outputs was abstracted into its own function so that it will be easier to attenuate long, droning signals as is planned.